+++ lib/core.ps1
@@ -24,7 +24,7 @@ function Get-Encoding($wc) {
 }
 
 function Get-UserAgent() {
-    return "Scoop/1.0 (+http://scoop.sh/) PowerShell/$($PSVersionTable.PSVersion.Major).$($PSVersionTable.PSVersion.Minor) (Windows NT $([System.Environment]::OSVersion.Version.Major).$([System.Environment]::OSVersion.Version.Minor); $(if($env:PROCESSOR_ARCHITECTURE -eq 'AMD64'){'Win64; x64; '})$(if($env:PROCESSOR_ARCHITEW6432 -eq 'AMD64'){'WOW64; '})$PSEdition)"
+    return "Scoop/1.0 (+http://scoop.sh/) PowerShell/$($PSVersionTable.PSVersion.Major).$($PSVersionTable.PSVersion.Minor) (Windows NT $([System.Environment]::OSVersion.Version.Major).$([System.Environment]::OSVersion.Version.Minor); $(if(${env:ProgramFiles(Arm)}){'ARM64; '}elseif($env:PROCESSOR_ARCHITECTURE -eq 'AMD64'){'Win64; x64; '})$(if($env:PROCESSOR_ARCHITEW6432 -in 'AMD64','ARM64'){'WOW64; '})$PSEdition)"
 }
 
 function Show-DeprecatedWarning {
@@ -857,15 +857,38 @@ function ensure_in_path($dir, $global) {
     }
 }
 
-function ensure_architecture($architecture_opt) {
-    if(!$architecture_opt) {
-        return default_architecture
+function Get-DefaultArchitecture {
+    $arch = get_config DEFAULT_ARCHITECTURE
+    $system = if (${env:ProgramFiles(Arm)}) {
+        'arm64'
+    } elseif ([System.Environment]::Is64BitOperatingSystem) {
+        '64bit'
+    } else {
+        '32bit'
     }
-    $architecture_opt = $architecture_opt.ToString().ToLower()
-    switch($architecture_opt) {
-        { @('64bit', '64', 'x64', 'amd64', 'x86_64', 'x86-64')  -contains $_ } { return '64bit' }
-        { @('32bit', '32', 'x86', 'i386', '386', 'i686')  -contains $_ } { return '32bit' }
-        default { throw [System.ArgumentException] "Invalid architecture: '$architecture_opt'"}
+    if ($null -eq $arch) {
+        $arch = $system
+    } else {
+        try {
+            $arch = Format-ArchitectureString $arch
+        } catch {
+            warn 'Invalid default architecture configured. Determining default system architecture'
+            $arch = $system
+        }
+    }
+    return $arch
+}
+
+function Format-ArchitectureString($Architecture) {
+    if (!$Architecture) {
+        return Get-DefaultArchitecture
+    }
+    $Architecture = $Architecture.ToString().ToLower()
+    switch ($Architecture) {
+        { @('64bit', '64', 'x64', 'amd64', 'x86_64', 'x86-64') -contains $_ } { return '64bit' }
+        { @('32bit', '32', 'x86', 'i386', '386', 'i686') -contains $_ } { return '32bit' }
+        { @('arm64', 'arm', 'aarch64') -contains $_ } { return 'arm64' }
+        default { throw [System.ArgumentException] "Invalid architecture: '$Architecture'" }
     }
 }
 
@@ -1239,6 +1262,9 @@ $globaldir = $env:SCOOP_GLOBAL, (get_config GLOBAL_PATH), "$env:ProgramData\scoo
 #       multiple users write and access cached files at the same time.
 #       Use at your own risk.
 $cachedir = $env:SCOOP_CACHE, (get_config CACHE_PATH), "$scoopdir\cache" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
+
+# OS information
+$WindowsBuild = [System.Environment]::OSVersion.Version.Build
 
 # Setup proxy globally
 setup_proxy