+++ libexec/scoop-download.ps1
@@ -14,15 +14,15 @@
 #      scoop download path\to\app.json
 #
 # Options:
-#   -f, --force               Force download (overwrite cache)
-#   -h, --no-hash-check       Skip hash verification (use with caution!)
-#   -u, --no-update-scoop     Don't update Scoop before downloading if it's outdated
-#   -a, --arch <32bit|64bit>  Use the specified architecture, if the app supports it
+#   -f, --force                     Force download (overwrite cache)
+#   -h, --no-hash-check             Skip hash verification (use with caution!)
+#   -u, --no-update-scoop           Don't update Scoop before downloading if it's outdated
+#   -a, --arch <32bit|64bit|arm64>  Use the specified architecture, if the app supports it
 
 . "$PSScriptRoot\..\lib\getopt.ps1"
 . "$PSScriptRoot\..\lib\json.ps1" # 'autoupdate.ps1' (indirectly)
 . "$PSScriptRoot\..\lib\autoupdate.ps1" # 'generate_user_manifest' (indirectly)
-. "$PSScriptRoot\..\lib\manifest.ps1" # 'default_architecture' 'generate_user_manifest' 'Get-Manifest'
+. "$PSScriptRoot\..\lib\manifest.ps1" # 'generate_user_manifest' 'Get-Manifest'
 . "$PSScriptRoot\..\lib\install.ps1"
 
 $opt, $apps, $err = getopt $args 'fhua:' 'force', 'no-hash-check', 'no-update-scoop', 'arch='
@@ -30,9 +30,9 @@ if ($err) { error "scoop download: $err"; exit 1 }
 
 $check_hash = !($opt.h -or $opt.'no-hash-check')
 $use_cache = !($opt.f -or $opt.force)
-$architecture = default_architecture
+$architecture = Get-DefaultArchitecture
 try {
-    $architecture = ensure_architecture ($opt.a + $opt.arch)
+    $architecture = Format-ArchitectureString ($opt.a + $opt.arch)
 } catch {
     abort "ERROR: $_"
 }
@@ -89,8 +89,9 @@ foreach ($curr_app in $apps) {
         $curr_check_hash = $false
     }
 
-    if(!(supports_architecture $manifest $architecture)) {
-        error "'$app' doesn't support $architecture architecture!"
+    $architecture = Get-SupportedArchitecture $manifest $architecture
+    if ($null -eq $architecture) {
+        error "'$app' doesn't support current architecture!"
         continue
     }